Address API review comments for extensions packages
