--- sdk/extensions/Azure.AspNetCore.DataProtection.Blobs/tests/Azure.AspNetCore.DataProtection.Blobs.Tests.csproj
@@ -1,23 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>$(RequiredTargetFrameworks)</TargetFrameworks>
-    <ExcludeRecordingFramework>true</ExcludeRecordingFramework>
-  </PropertyGroup>
-
-  <Import Project="..\..\..\core\Azure.Core\tests\TestFramework.props" />
-
-  <ItemGroup>
-    <PackageReference Include="Azure.Identity" />
-    <PackageReference Include="nunit" />
-    <PackageReference Include="NUnit3TestAdapter" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
-    <PackageReference Include="Moq" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\src\Azure.AspNetCore.DataProtection.Blobs.csproj" />
-  </ItemGroup>
-
-</Project>
