--- sdk/extensions/Azure.AspNetCore.DataProtection.Keys/CHANGELOG.md
@@ -1,8 +1,0 @@
-# Release History
-
-## 1.0.0-preview.2 (Unreleased)
-
-
-## 1.0.0-preview.1 (2020)
-
-- Initial preview of the Azure.AspNetCore.DataProtection.Keys library
