--- sdk/extensions/Azure.AspNetCore.DataProtection.Keys/api/Azure.AspNetCore.DataProtection.Keys.netstandard2.0.cs
@@ -1,8 +1,0 @@
-namespace Microsoft.AspNetCore.DataProtection
-{
-    public static partial class AzureDataProtectionKeyVaultKeyBuilderExtensions
-    {
-        public static Microsoft.AspNetCore.DataProtection.IDataProtectionBuilder ProtectKeysWithAzureKeyVault(this Microsoft.AspNetCore.DataProtection.IDataProtectionBuilder builder, string keyIdentifier, Azure.Core.Cryptography.IKeyEncryptionKeyResolver keyResolver) { throw null; }
-        public static Microsoft.AspNetCore.DataProtection.IDataProtectionBuilder ProtectKeysWithAzureKeyVault(this Microsoft.AspNetCore.DataProtection.IDataProtectionBuilder builder, string keyIdentifier, Azure.Core.TokenCredential tokenCredential) { throw null; }
-    }
-}
