--- sdk/extensions/Azure.AspNetCore.DataProtection.Keys/src/Azure.AspNetCore.DataProtection.Keys.csproj
@@ -1,21 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <Description>Microsoft Azure Key Vault key encryption support.</Description>
-    <PackageTags>aspnetcore;dataprotection;azure;keyvault</PackageTags>
-    <Version>1.0.0-preview.2</Version>
-    <EnableApiCompat>false</EnableApiCompat>
-    <IsExtensionClientLibrary>true</IsExtensionClientLibrary>
-    <NoWarn>$(NoWarn);AZC0102</NoWarn>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.DataProtection" />
-    <PackageReference Include="Azure.Security.KeyVault.Keys" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Include="$(AzureCoreSharedSources)Argument.cs" />
-  </ItemGroup>
-
-</Project>
