--- sdk/extensions/Azure.AspNetCore.DataProtection.Keys/tests/Azure.AspNetCore.DataProtection.Keys.Tests.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>$(RequiredTargetFrameworks)</TargetFrameworks>
-    <ExcludeRecordingFramework>true</ExcludeRecordingFramework>
-  </PropertyGroup>
-
-  <Import Project="..\..\..\core\Azure.Core\tests\TestFramework.props" />
-
-  <ItemGroup>
-    <PackageReference Include="Azure.Identity" />
-    <PackageReference Include="nunit" />
-    <PackageReference Include="NUnit3TestAdapter" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
-    <PackageReference Include="Moq" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\src\Azure.AspNetCore.DataProtection.Keys.csproj" />
-  </ItemGroup>
-
-</Project>
