--- sdk/extensions/Azure.AspNetCore.DataProtection.Keys/tests/Snippets.cs
@@ -1,25 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System;
-using Azure.Identity;
-using Microsoft.AspNetCore.DataProtection;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Azure.AspNetCore.DataProtection.Keys.Tests
-{
-    public class Snippets
-    {
-        private class StartupIdentity
-        {
-            #region Snippet:ProtectKeysWithAzureKeyVault
-            public void ConfigureServices(IServiceCollection services)
-            {
-                services
-                    .AddDataProtection()
-                    .ProtectKeysWithAzureKeyVault("<Key-ID>", new DefaultAzureCredential());
-            }
-            #endregion
-        }
-    }
-}
