+++ sdk/extensions/Azure.Extensions.AspNetCore.Configuration.Secrets/CHANGELOG.md
