+++ sdk/extensions/Azure.Extensions.AspNetCore.Configuration.Secrets/tests/Azure.Extensions.AspNetCore.Configuration.Secrets.Tests.csproj
@@ -1,0 +1,23 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>$(RequiredTargetFrameworks)</TargetFrameworks>
+    <ExcludeRecordingFramework>true</ExcludeRecordingFramework>
+  </PropertyGroup>
+
+  <Import Project="..\..\..\core\Azure.Core\tests\TestFramework.props" />
+
+  <ItemGroup>
+    <PackageReference Include="Azure.Identity" />
+    <PackageReference Include="nunit" />
+    <PackageReference Include="NUnit3TestAdapter" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" />
+    <PackageReference Include="Microsoft.Extensions.Configuration" />
+    <PackageReference Include="Moq" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\src\Azure.Extensions.AspNetCore.Configuration.Secrets.csproj" />
+  </ItemGroup>
+
+</Project>
