--- sdk/extensions/Azure.Extensions.Configuration.Secrets/tests/ConfigurationProviderExtensions.cs
@@ -1,23 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System;
-using Microsoft.Extensions.Configuration;
-
-namespace Azure.Extensions.Configuration.Secrets.Tests
-{
-    public static class ConfigurationProviderExtensions
-    {
-        public static string Get(this IConfigurationProvider provider, string key)
-        {
-            string value;
-
-            if (!provider.TryGet(key, out value))
-            {
-                throw new InvalidOperationException("Key not found");
-            }
-
-            return value;
-        }
-    }
-}
