+++ sdk/extensions/Azure.Extensions.sln
@@ -3,23 +3,23 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Azure.AspNetCore.DataProtection.Blobs", "Azure.AspNetCore.DataProtection.Blobs", "{12C39FF8-57E3-4A24-AE05-42A104EC9C69}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Azure.Extensions.AspNetCore.DataProtection.Blobs", "Azure.Extensions.AspNetCore.DataProtection.Blobs", "{12C39FF8-57E3-4A24-AE05-42A104EC9C69}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.AspNetCore.DataProtection.Blobs", "Azure.AspNetCore.DataProtection.Blobs\src\Azure.AspNetCore.DataProtection.Blobs.csproj", "{70DAB56B-C642-480D-9F47-D490959A3431}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Extensions.AspNetCore.DataProtection.Blobs", "Azure.Extensions.AspNetCore.DataProtection.Blobs\src\Azure.Extensions.AspNetCore.DataProtection.Blobs.csproj", "{70DAB56B-C642-480D-9F47-D490959A3431}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.AspNetCore.DataProtection.Blobs.Tests", "Azure.AspNetCore.DataProtection.Blobs\tests\Azure.AspNetCore.DataProtection.Blobs.Tests.csproj", "{CE0F89E2-9894-43DE-A9F6-90C5D186DD84}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Extensions.AspNetCore.DataProtection.Blobs.Tests", "Azure.Extensions.AspNetCore.DataProtection.Blobs\tests\Azure.Extensions.AspNetCore.DataProtection.Blobs.Tests.csproj", "{CE0F89E2-9894-43DE-A9F6-90C5D186DD84}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Azure.AspNetCore.DataProtection.Keys", "Azure.AspNetCore.DataProtection.Keys", "{B358FBE9-0833-484D-B988-588B75AE99CF}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Azure.Extensions.AspNetCore.DataProtection.Keys", "Azure.Extensions.AspNetCore.DataProtection.Keys", "{B358FBE9-0833-484D-B988-588B75AE99CF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.AspNetCore.DataProtection.Keys", "Azure.AspNetCore.DataProtection.Keys\src\Azure.AspNetCore.DataProtection.Keys.csproj", "{4324C3A5-67DA-4130-B506-69BDB282DB42}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Extensions.AspNetCore.DataProtection.Keys", "Azure.Extensions.AspNetCore.DataProtection.Keys\src\Azure.Extensions.AspNetCore.DataProtection.Keys.csproj", "{4324C3A5-67DA-4130-B506-69BDB282DB42}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.AspNetCore.DataProtection.Keys.Tests", "Azure.AspNetCore.DataProtection.Keys\tests\Azure.AspNetCore.DataProtection.Keys.Tests.csproj", "{0BE93DC0-97FC-48D6-BA9F-482BCF974D83}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Extensions.AspNetCore.DataProtection.Keys.Tests", "Azure.Extensions.AspNetCore.DataProtection.Keys\tests\Azure.Extensions.AspNetCore.DataProtection.Keys.Tests.csproj", "{0BE93DC0-97FC-48D6-BA9F-482BCF974D83}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Azure.Extensions.Configuration.Secrets", "Azure.Extensions.Configuration.Secrets", "{A6B68F68-8FDF-4D1F-9277-BABEE6904DCA}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Azure.Extensions.AspNetCore.Configuration.Secrets", "Azure.Extensions.AspNetCore.Configuration.Secrets", "{A6B68F68-8FDF-4D1F-9277-BABEE6904DCA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Extensions.Configuration.Secrets", "Azure.Extensions.Configuration.Secrets\src\Azure.Extensions.Configuration.Secrets.csproj", "{8EFE342E-7C53-4489-B14B-EB813CE015B6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Extensions.AspNetCore.Configuration.Secrets", "Azure.Extensions.AspNetCore.Configuration.Secrets\src\Azure.Extensions.AspNetCore.Configuration.Secrets.csproj", "{8EFE342E-7C53-4489-B14B-EB813CE015B6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Extensions.Configuration.Secrets.Tests", "Azure.Extensions.Configuration.Secrets\tests\Azure.Extensions.Configuration.Secrets.Tests.csproj", "{5F56D5D1-670D-4C34-BB1A-E4D46F0CD05E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Extensions.AspNetCore.Configuration.Secrets.Tests", "Azure.Extensions.AspNetCore.Configuration.Secrets\tests\Azure.Extensions.AspNetCore.Configuration.Secrets.Tests.csproj", "{5F56D5D1-670D-4C34-BB1A-E4D46F0CD05E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
